[ENGOPS-1125] Refactor package references from emr32 to emr34
