+++ emr34/src/META-INF/services/org.pentaho.hadoop.shim.spi.HadoopShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.authorization.AuthenticatingHadoopShim
+org.pentaho.hadoop.shim.emr34.authorization.AuthenticatingHadoopShim
