+++ emr34/src/META-INF/services/org.pentaho.hadoop.shim.spi.PigShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.delegating.DelegatingPigShim
+org.pentaho.hadoop.shim.emr34.delegating.DelegatingPigShim
