+++ emr34/src/META-INF/services/org.pentaho.hadoop.shim.spi.SnappyShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.delegating.DelegatingSnappyShim
+org.pentaho.hadoop.shim.emr34.delegating.DelegatingSnappyShim
