+++ emr34/src/META-INF/services/org.pentaho.hadoop.shim.spi.SqoopShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.delegating.DelegatingSqoopShim
+org.pentaho.hadoop.shim.emr34.delegating.DelegatingSqoopShim
