+++ emr34/src/META-INF/services/org.pentaho.hbase.shim.spi.HBaseShim
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.delegating.DelegatingHBaseShim
+org.pentaho.hadoop.shim.emr34.delegating.DelegatingHBaseShim
