+++ emr34/src/META-INF/services/org.pentaho.oozie.shim.api.OozieClientFactory
@@ -1,1 +1,1 @@
-org.pentaho.hadoop.shim.emr32.delegating.DelegatingOozieClientFactory
+org.pentaho.hadoop.shim.emr34.delegating.DelegatingOozieClientFactory
