+++ emr34/src/org/pentaho/hadoop/shim/emr32/ClassPathModifyingSqoopShim.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.hadoop.shim.emr32;
+package org.pentaho.hadoop.shim.emr34;
 
 import java.util.concurrent.Callable;
