+++ emr34/src/org/pentaho/hadoop/shim/emr32/PigShim.java
@@ -19,7 +19,7 @@
  * limitations under the License.
  *
  ******************************************************************************/
-package org.pentaho.hadoop.shim.emr32;
+package org.pentaho.hadoop.shim.emr34;
 
 import org.apache.pig.PigServer;
 import org.apache.pig.tools.grunt.GruntParser;
