+++ emr34/src/org/pentaho/hadoop/shim/emr32/delegating/DelegatingSqoopShim.java
@@ -20,12 +20,12 @@
 *
 ******************************************************************************/
 
-package org.pentaho.hadoop.shim.emr32.delegating;
+package org.pentaho.hadoop.shim.emr34.delegating;
 
 import org.pentaho.hadoop.shim.ShimVersion;
 import org.pentaho.hadoop.shim.api.Configuration;
-import org.pentaho.hadoop.shim.emr32.authorization.HadoopAuthorizationService;
-import org.pentaho.hadoop.shim.emr32.authorization.HasHadoopAuthorizationService;
+import org.pentaho.hadoop.shim.emr34.authorization.HadoopAuthorizationService;
+import org.pentaho.hadoop.shim.emr34.authorization.HasHadoopAuthorizationService;
 import org.pentaho.hadoop.shim.spi.SqoopShim;
 
 public class DelegatingSqoopShim implements SqoopShim, HasHadoopAuthorizationService {
